Add variant functionality and remove titles / multiple questions
This is still something of a work in progress. I think the main
functionality is there, but it needs quite a lot of tests.

The test_variant_questions.json should show the main functionality of
question_variants, but needs to be expanded to cover more scenarios.

This PR deletes quite a lot of code which was there to handle multiple
questions and titles. I have commented out some of the functional tests
to get things working, but those tests need to be rejigged to
accomodate the changes. Currently all of the unit tests are passing and
the schemas should all pass validation.

Coverage is not 100% at the moment, though most of this will be covered
by the integration tests.